Enhanced physics collisions with the ground. Fixed ~30 warnings in visual studio (stupid double->float)

+++ Sample/source/ActorSystem.cpp
@@ -68,7 +68,7 @@ void ActorSystem::Update(double delta, double now) {
 	for (unsigned int i = 0; i <= highestFull; i++) {
 		if (allActors[i] != NULL) {
 			newFullSize = i;
-			allActors[i]->Update(delta,now);
+			allActors[i]->Update((float)delta,(float)now);
 		}
 	}
